Style: brace placement and cast fixes in storedevice
Apply K&R brace style (opening brace on same line) and add explicit static_cast<unsigned> for GetSize() calls in printf to avoid potential type mismatch warnings.

+++ lib-configstore/device/i2c/storedevice.cpp
@@ -39,8 +39,7 @@
 #include "i2c/at24cxx.h" // IWYU pragma: keep
 #include "firmware/debug/debug_debug.h"
 
-namespace storedevice
-{
+namespace storedevice {
 #if !defined(CONFIG_FLASHROM_I2C_INDEX)
 #define CONFIG_FLASHROM_I2C_INDEX 0
 #endif
@@ -50,42 +49,34 @@ static constexpr auto kFlashSectorSize = 4096U;
 static constexpr auto kRomSize = 4096U;
 } // namespace storedevice
 
-StoreDevice::StoreDevice() : AT24C32(storedevice::kI2CIndex)
-{
+StoreDevice::StoreDevice() : AT24C32(storedevice::kI2CIndex) {
     DEBUG_ENTRY();
 
     detected_ = AT24C32::IsConnected();
 
-    if (!detected_)
-    {
+    if (!detected_) {
         printf("StoreDevice: No AT24C32 at %2x", AT24C32::GetAddress());
-    }
-    else
-    {
-        printf("StoreDevice: AT24C32 total %u bytes [%u kB]\n", GetSize(), GetSize() / 1024U);
+    } else {
+        printf("StoreDevice: AT24C32 total %u bytes [%u kB]\n", static_cast<unsigned>(GetSize()), static_cast<unsigned>(GetSize() / 1024U));
     }
 
     DEBUG_EXIT();
 }
 
-StoreDevice::~StoreDevice()
-{
+StoreDevice::~StoreDevice() {
     DEBUG_ENTRY();
     DEBUG_EXIT();
 }
 
-uint32_t StoreDevice::GetSize() const
-{
+uint32_t StoreDevice::GetSize() const {
     return storedevice::kRomSize;
 }
 
-uint32_t StoreDevice::GetSectorSize() const
-{
+uint32_t StoreDevice::GetSectorSize() const {
     return storedevice::kFlashSectorSize;
 }
 
-bool StoreDevice::Read(uint32_t offset, uint32_t length, uint8_t* buffer, storedevice::Result& result)
-{
+bool StoreDevice::Read(uint32_t offset, uint32_t length, uint8_t* buffer, storedevice::Result& result) {
     DEBUG_ENTRY();
     assert((offset + length) <= storedevice::ROM_SIZE);
 
@@ -97,8 +88,7 @@ bool StoreDevice::Read(uint32_t offset, uint32_t length, uint8_t* buffer, stored
     return true;
 }
 
-bool StoreDevice::Erase([[maybe_unused]] uint32_t offset, [[maybe_unused]] uint32_t length, storedevice::Result& result)
-{
+bool StoreDevice::Erase([[maybe_unused]] uint32_t offset, [[maybe_unused]] uint32_t length, storedevice::Result& result) {
     DEBUG_ENTRY();
 
     result = storedevice::Result::kOk;
@@ -107,8 +97,7 @@ bool StoreDevice::Erase([[maybe_unused]] uint32_t offset, [[maybe_unused]] uint3
     return true;
 }
 
-bool StoreDevice::Write(uint32_t offset, uint32_t length, const uint8_t* buffer, storedevice::Result& result)
-{
+bool StoreDevice::Write(uint32_t offset, uint32_t length, const uint8_t* buffer, storedevice::Result& result) {
     DEBUG_ENTRY();
     assert((offset + length) <= ROM_SIZE);
